requirements.txt
faiss-cpu~=1.11.0
langchain~=0.3.25
feedparser~=6.0.11
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
fastapi~=0.115.9
pydantic~=2.10.6